backend/requirements.txt
# Web Framework
fastapi==0.111.1
uvicorn[standard]==0.30.1
python-multipart==0.0.9

# Database
sqlalchemy==2.0.31
alembic==1.13.2
psycopg2-binary==2.9.9
asyncpg==0.29.0

# Pydantic for data validation
pydantic==2.8.2
pydantic-settings==2.3.4

# AI and Machine Learning
openai==1.37.1
langchain==0.2.11
langchain-community==0.2.10
langchain-openai==0.1.19
transformers==4.43.3
torch==2.4.0
sentence-transformers==3.0.1
numpy==1.26.4
scikit-learn==1.5.1

# IBM Watsonx (optional, for production)
# ibm-watsonx-ai==1.1.2  # Skip for demo
# ibm-watson-machine-learning==1.0.362  # Skip for demo

# Hugging Face
huggingface-hub==0.24.2
datasets==2.20.0

# Replicate (for IBM Granite via Replicate)
replicate==0.25.1

# HTTP Client
httpx==0.27.0
aiohttp==3.9.5
requests==2.32.3

# Environment and Configuration
python-dotenv==1.0.1
pydantic-settings==2.3.4

# CORS
fastapi-cors==0.0.6

# Authentication (optional)
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9

# Caching
redis==5.0.7
aioredis==2.0.1

# Logging and Monitoring
loguru==0.7.2
prometheus-client==0.20.0

# Data Processing
pandas==2.2.2
geopy==2.4.1
googlemaps==4.10.0

# Testing
pytest==8.3.2
pytest-asyncio==0.23.8
httpx==0.27.0

# Development
black==24.4.2
isort==5.13.2
flake8==7.1.0
mypy==1.11.0

# Utilities
python-slugify==8.0.4
Pillow==10.4.0
